Разделил js код на несколько модулей
